bug fix after review
