--- packages/client/src/hooks/useOAuth.ts
@@ -1,5 +1,0 @@
-export const useOAuth = (queryString: string) => {
-  const urlParams = new URLSearchParams(queryString);
-  const code = urlParams.get('code');
-  return code;
-};
